backend/requirements.txt
fastapi
uvicorn
python-dotenv
airtable-python-wrapper
openai
pydantic
requests
python-dateutil
PyMuPDF 